updating classes, + pwd for user

+++ src/main/java/com/zappts/CRUDMTG/config/validation/CardDoesNotExistException.java
@@ -2,7 +2,7 @@ package com.zappts.CRUDMTG.config.validation;
 
 public class CardDoesNotExistException extends RuntimeException{
 	  
-	public CardDoesNotExistException(Integer idcard){
+	public CardDoesNotExistException(Long idcard){
 		    super("Carta " + idcard + " não existe!");
 		  }
