+++ src/main/java/com/zappts/CRUDMTG/config/validation/PlayerDoesNotExistException.java
@@ -2,7 +2,7 @@ package com.zappts.CRUDMTG.config.validation;
 
 public class PlayerDoesNotExistException extends RuntimeException{
 	  
-	public PlayerDoesNotExistException(Integer idplayer){
+	public PlayerDoesNotExistException(Long idplayer){
 		    super("Jogador " + idplayer + " não existe!");
 		  }
