+++ src/main/java/com/zappts/CRUDMTG/repository/ListsRepository.java
@@ -1,12 +1,10 @@
 package com.zappts.CRUDMTG.repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.zappts.CRUDMTG.model.Lists;
 
-public interface ListsRepository extends JpaRepository<Lists, Integer> {
+public interface ListsRepository extends JpaRepository<Lists, Long> {
 
 //	 public List<Lists> findAll();
 }
